added types, updated demo to include conditions
